order_fulfiller/backend/requirements.txt
Django==3.2.2
django-cors-headers==3.7.0
django-decouple==2.0
django-sslserver==0.22
djangorestframework==3.12.4
mysqlclient==2.0.3
requests==2.25.1